Implement 'for' loops.
But oh, they're inefficient! Lots of jumping around because we emit code in the order it gets parsed, not in the order we need it. But I think it'll work! When I switch to generating a parse tree, and translating that, we can generate code in the right order and avoid the extraneous jumps. TODO: allow for empty expressions, e.g. for (;;) meaning loop forever.