Add test code with function calls.
'simplefunc.c' is a kludgy way to call UK109 monitor ROM functions from C. But it might just work! This is the first C test program that stands a chance of actually working on real 6809 hardware in a way that we can see visible results from. I'll run it on the UK109 as soon as I can.