Begin to implement types.
Now that we have the beginnings of a symbol table, we can emit the correct code to load, store, and increment scalar variables of different types. Only 'int' and 'char' really work as yet, because they're both loaded into the D register. We can't parse 'long' yet. 'float' and 'double' are parsed but we can't do much with the 6309's Q register yet, and we certainly don't have a floating-point library. Incrementing or decrementing a 'char' variable is actually reasonably efficient now. TODO: emit storage for 'extern' and 'static' objects via a pointer to the symbol table too.